Moved the Bert scanner to `joe.scanner`
