+++ lib/src/main/java/com/wjduquette/joe/bert/ErrorReporter.java
@@ -2,6 +2,6 @@ package com.wjduquette.joe.bert;
 
 import com.wjduquette.joe.Trace;
 
-interface ErrorReporter {
+public interface ErrorReporter {
     void reportError(Trace trace, boolean incomplete);
 }
